Updated modules/modis-prep.js (added 3 lines, deleted 1 line).

+++ modules/modis-prep.js
@@ -2,6 +2,8 @@
 
 
 var rescale = function(img) {
-  return img.multiply(0.0001).float().copyProperties(img);
+  return img.multiply(0.0001).copyProperties(img);
 };
 
+
+//float?
